+++ codebase/ext/dhtmlxscheduler_collision.js
@@ -1,15 +1,15 @@
 /*
 
 @license
-dhtmlxScheduler v.5.3.4 Stardard
+dhtmlxScheduler v.5.3.5 Stardard
 
 To use dhtmlxScheduler in non-GPL projects (and get Pro version of the product), please obtain Commercial/Enterprise or Ultimate license on our site https://dhtmlx.com/docs/products/dhtmlxScheduler/#licensing or contact us at sales@dhtmlx.com
 
 (c) XB Software Ltd.
 
 */
-Scheduler.plugin(function(t){!function(){function e(e){t._get_section_view()&&e&&(i=t.getEvent(e)[t._get_section_property()])}var i,a;t.config.collision_limit=1,t.attachEvent("onBeforeDrag",function(t){return e(t),!0}),t.attachEvent("onBeforeLightbox",function(i){var n=t.getEvent(i);return a=[n.start_date,n.end_date],e(i),!0}),t.attachEvent("onEventChanged",function(e){if(!e||!t.getEvent(e))return!0;var i=t.getEvent(e);if(!t.checkCollision(i)){if(!a)return!1;i.start_date=a[0],i.end_date=a[1],
-i._timed=this.isOneDayEvent(i)}return!0}),t.attachEvent("onBeforeEventChanged",function(e,i,a){return t.checkCollision(e)}),t.attachEvent("onEventAdded",function(e,i){t.checkCollision(i)||t.deleteEvent(e)}),t.attachEvent("onEventSave",function(e,i,a){if(i=t._lame_clone(i),i.id=e,!i.start_date||!i.end_date){var n=t.getEvent(e);i.start_date=new Date(n.start_date),i.end_date=new Date(n.end_date)}return i.rec_type&&t._roll_back_dates(i),t.checkCollision(i)}),
-t._check_sections_collision=function(e,i){var a=t._get_section_property();return e[a]==i[a]&&e.id!=i.id},t.checkCollision=function(e){var a=[],n=t.config.collision_limit;if(e.rec_type)for(var r=t.getRecDates(e),s=0;s<r.length;s++)for(var o=t.getEvents(r[s].start_date,r[s].end_date),d=0;d<o.length;d++)(o[d].event_pid||o[d].id)!=e.id&&a.push(o[d]);else{a=t.getEvents(e.start_date,e.end_date);for(var _=0;_<a.length;_++){var l=a[_]
-;if(l.id==e.id||l.event_length&&[l.event_pid,l.event_length].join("#")==e.id){a.splice(_,1);break}}}var h=t._get_section_view(),c=t._get_section_property(),u=!0;if(h){for(var g=0,_=0;_<a.length;_++)a[_].id!=e.id&&this._check_sections_collision(a[_],e)&&g++;g>=n&&(u=!1)}else a.length>=n&&(u=!1);if(!u){var f=!t.callEvent("onEventCollision",[e,a]);return f||(e[c]=i||e[c]),f}return u}}()});
+Scheduler.plugin(function(e){!function(){function t(t){e._get_section_view()&&t&&(i=e.getEvent(t)[e._get_section_property()])}var i,a;e.config.collision_limit=1,e.attachEvent("onBeforeDrag",function(e){return t(e),!0}),e.attachEvent("onBeforeLightbox",function(i){var n=e.getEvent(i);return a=[n.start_date,n.end_date],t(i),!0}),e.attachEvent("onEventChanged",function(t){if(!t||!e.getEvent(t))return!0;var i=e.getEvent(t);if(!e.checkCollision(i)){if(!a)return!1;i.start_date=a[0],i.end_date=a[1],
+i._timed=this.isOneDayEvent(i)}return!0}),e.attachEvent("onBeforeEventChanged",function(t,i,a){return e.checkCollision(t)}),e.attachEvent("onEventAdded",function(t,i){e.checkCollision(i)||e.deleteEvent(t)}),e.attachEvent("onEventSave",function(t,i,a){if(i=e._lame_clone(i),i.id=t,!i.start_date||!i.end_date){var n=e.getEvent(t);i.start_date=new Date(n.start_date),i.end_date=new Date(n.end_date)}return i.rec_type&&e._roll_back_dates(i),e.checkCollision(i)}),
+e._check_sections_collision=function(t,i){var a=e._get_section_property();return t[a]==i[a]&&t.id!=i.id},e.checkCollision=function(t){var a=[],n=e.config.collision_limit;if(t.rec_type)for(var r=e.getRecDates(t),s=0;s<r.length;s++)for(var o=e.getEvents(r[s].start_date,r[s].end_date),d=0;d<o.length;d++)(o[d].event_pid||o[d].id)!=t.id&&a.push(o[d]);else{a=e.getEvents(t.start_date,t.end_date);for(var _=0;_<a.length;_++){var l=a[_]
+;if(l.id==t.id||l.event_length&&[l.event_pid,l.event_length].join("#")==t.id){a.splice(_,1);break}}}var h=e._get_section_view(),c=e._get_section_property(),u=!0;if(h){for(var g=0,_=0;_<a.length;_++)a[_].id!=t.id&&this._check_sections_collision(a[_],t)&&g++;g>=n&&(u=!1)}else a.length>=n&&(u=!1);if(!u){var f=!e.callEvent("onEventCollision",[t,a]);return f||(t[c]=i||t[c]),f}return u}}()});
 //# sourceMappingURL=../sources/ext/dhtmlxscheduler_collision.js.map