+++ codebase/sources/ext/dhtmlxscheduler_layer.js
@@ -1,7 +1,7 @@
 /*
 
 @license
-dhtmlxScheduler v.5.3.4 Stardard
+dhtmlxScheduler v.5.3.5 Stardard
 
 To use dhtmlxScheduler in non-GPL projects (and get Pro version of the product), please obtain Commercial/Enterprise or Ultimate license on our site https://dhtmlx.com/docs/products/dhtmlxScheduler/#licensing or contact us at sales@dhtmlx.com
 
@@ -286,7 +286,7 @@ scheduler.attachEvent("onTemplatesReady",function(){
 		var cse = (scheduler.templates['event_class_'+ev.layer])?scheduler.templates['event_class_'+ev.layer](ev.start_date,ev.end_date,ev):scheduler.templates.event_class(ev.start_date,ev.end_date,ev);
 		if (cse) cs=cs+" "+cse;
 
-		var borderBox = scheduler._border_box_bvents();
+		var borderBox = scheduler._border_box_events();
 
 		var borderBoxWidth = w - 2;
 		var boxWidth = borderBox ? borderBoxWidth : (w-4),